requirements.txt
python
django
djangorestframework
requests
